pyproject.toml
[project]
name = "email-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.0",
    "aiohttp>=3.12.15",
    "fastapi==0.104.1",
    "httpx==0.25.2",
    "msal==1.24.1",
    "openai==1.3.8",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "typing-extensions==4.8.0",
    "uvicorn[standard]==0.24.0",
]
